+++ Assets/Plugins/CozyColdEnvironments.Unity.ECS/Scripts/Items/_buffer/InventoryRemoveItemQuery.cs
@@ -63,7 +63,7 @@ namespace CCEnvs.UnityX.ECS.Items
         [BurstCompile]
         public static NativeArray<InventoryRemoveItemQuery> GetInventoryQueries(
             this NativeArray<InventoryRemoveItemQuery> queries,
-            int inventoryID,
+            long inventoryID,
             Allocator allocator
             )
         {
@@ -91,7 +91,7 @@ namespace CCEnvs.UnityX.ECS.Items
         [BurstCompile]
         public static int GetInventoryQueryCount(
             this NativeArray<InventoryRemoveItemQuery> queries,
-            int inventoryID
+            long inventoryID
             )
         {
             int count = 0;
@@ -106,7 +106,7 @@ namespace CCEnvs.UnityX.ECS.Items
         [BurstCompile]
         public static bool HasInventoryQuery(
             this NativeArray<InventoryRemoveItemQuery> queries,
-            int inventoryID
+            long inventoryID
             )
         {
             for (int i = 0; i < queries.Length; i++)
